vendai_desktop/requirements.txt
pywebview
psutil
requests
# Optional: PyQt6 if you prefer the Qt approach
# PyQt6
# PyQt6-WebEngine
